// C1021G1-Repo-FE-new/project-c1021g1-sprint1/src/app/ticket/model/TicketDTO.ts
export interface TicketDTO {

  idTicket: number;

  birthdayTicket: string;

  buyerTicket: string;

  codeTicket: string;

  delFlagTicket: boolean;

  emailTicket: string;

  genderTicket: boolean;

  phoneTicket: string;

  pointTicket: number;

  priceTicket: number;

  statusTicket: boolean;

  nameCustomer: string;

  nameEmployee: string;

  codeSeat: string;

  toFlight: string;

  fromFlight: string;

  dateStart: string;

  nameSeatType: string;

  dateTicket: string;
  idCardTicket: string;
}